requirements.txt
# WhatsApp API
requests==2.31.0
python-dotenv==1.0.0

# Discord Bot
discord.py==2.3.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Async Support
asyncio==3.4.3
aiohttp==3.8.6

# OpenAI for summarization
openai==1.3.5

# Background tasks and scheduling
apscheduler==3.10.4

# Logging and error handling
loguru==0.7.2

# CLI menu
simple-term-menu==1.6.1 